Noteworthy differences: add two MATLAB/Julia idioms
Briefly describe two MATLAB idioms from @dgleich, and their Julia
equivalents.

The first one is logical indexing of an array with a logical vector
involving the array itself, like in x(x>3) and x(x>3) = [].
List filter/filter! as alternatives to logical indexing.

The second one is to describe extracting or "dereferencing" a cell
array, an operation which in Julia uses splatting.

Also, capitalize MATLAB in the preferred way.

[ci skip]